pyproject.toml
[tool.poetry]
name = "padfoot_bot"
version = "0.1.0"
description = ""
authors = ["Costa"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.7.4.post0"
anyio = "3.6.1"
APScheduler = "3.9.1"
asgiref = "3.5.0"
async-timeout = "3.0.1"
attrs = "21.4.0"
beautifulsoup4 = "4.11.1"
certifi = "2022.6.15"
chardet = "4.0.0"
charset-normalizer = "2.1.0"
click = "8.1.3"
cloudscraper = "1.2.60"
"discord.py" = "1.7.3"
fastapi = "0.78.0"
h11 = "0.13.0"
httptools = "0.4.0"
idna = "3.3"
multidict = "6.0.2"
pydantic = "1.9.1"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
python-dotenv = "0.20.0"
pytz = "2022.1"
pytz-deprecation-shim = "0.1.0.post0"
PyYAML = "6.0"
requests = "2.28.1"
requests-toolbelt = "0.9.1"
six = "1.16.0"
sniffio = "1.2.0"
soupsieve = "2.3.2.post1"
starlette = "0.19.1"
typing-extensions = "4.2.0"
tzdata = "2022.1"
tzlocal = "4.2"
urllib3 = "1.26.9"
uvicorn = "0.18.2"
uvloop = "0.16.0"
watchfiles = "0.15.0"
watchgod = "0.8.1"
wcwidth = "0.2.5"
websockets = "10.3"
yarl = "1.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
